# swagger.yaml
swagger: '2.0'
info:
  description: |
    Application deployment and management should be automated, auditable, and easy to understand and that\'s what beetle tries to achieve in a simple manner. Beetle automates the deployment and rollback of your applications in a multi-cluster, multi-namespaces kubernetes environments. Easy to integrate with through API endpoints & webhooks to fit a variety of workflows.
  version: 1.0.0
  title: Beetle
  contact:
    email: hello@clivern.com
  license:
    name: MIT
    url: 'https://github.com/Clivern/Beetle/blob/main/LICENSE'
host: beetle.yourcompany.com
basePath: /
schemes:
  - https
  - http
paths:
...
  /api/v1/cluster:
    get:
      tags:
        - Cluster
      summary: Get clusters list
      description: ''
      operationId: getClusters
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Clusters'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '500':
          description: Internal server error
      security:
        - api_key: []
...
  '/api/v1/cluster/{cn}/namespace':
    get:
      tags:
        - Namespace
      summary: Get namespaces list
      description: ''
      operationId: getNamespaces
      produces:
        - application/json
      parameters:
        - in: path
          name: cn
          description: The name of the cluster
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Namespaces'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '500':
          description: Internal server error
      security:
        - api_key: []
...
  '/api/v1/job/{uuid}':
    get:
      tags:
        - Job
      summary: Get a job by UUID
      description: ''
      operationId: getJobByUUID
      produces:
        - application/json
      parameters:
        - in: path
          name: uuid
          description: The UUID of the job
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Job'
        '400':
          description: Invalid request
        '404':
          description: Job not found
        '500':
          description: Internal server error
      security:
        - api_key: []
    delete:
      tags:
        - Job
      summary: Delete a job by UUID
      description: ''
      operationId: deleteJobByUUID
      produces:
        - application/json
      parameters:
        - in: path
          name: uuid
          description: The UUID of the job
          required: true
          type: string
      responses:
        '204':
          description: successful operation
        '400':
          description: Invalid request
        '404':
          description: Job not found
        '500':
          description: Internal server error
      security:
        - api_key: []
securityDefinitions:
  api_key:
    type: apiKey
    name: X-API-KEY
    in: header
definitions:
...
  Cluster:
    type: object
    properties:
      name:
        type: string
      health:
        type: boolean
  Clusters:
    type: object
    properties:
      clusters:
        type: array
        items:
          $ref: '#/definitions/Cluster'
  Namespace:
    type: object
    properties:
      name:
        type: string
      uid:
        type: string
      status:
        type: string
  Namespaces:
    type: object
    properties:
      namespaces:
        type: array
        items:
          $ref: '#/definitions/Namespace'
  Job:
    type: object
    properties:
      id:
        type: integer
        format: int64
      uuid:
        type: string
      payload:
        type: string
      status:
        type: string
      type:
        type: string
      result:
        type: string
      retry:
        type: integer
        format: int64
      parent:
        type: integer
        format: int64
      run_at:
        type: string
      created_at:
        type: string
      updated_at:
        type: string
...